[Feat] 플레이리스트에서 직접 재생
